reactor_actor

Trait ActorProcess

pub trait ActorProcess: Send + 'static {
    type IMsg: Msg;
    type OMsg: Msg;

    // Required method
    fn process(&mut self, input: Self::IMsg) -> Vec<Self::OMsg>;
}

The ActorProcess trait defines the processing logic for an actor that transforms input messages into one or more output messages.

Example

struct Incrementer;

impl ActorProcess for Incrementer {
    type IMsg = i32;
    type OMsg = i32;

    fn process(&mut self, input: i32) -> Vec<i32> {
        vec![input + 1]
    }
}

Required Associated Types

type IMsg: Msg

The type of messages this actor accepts as input.

type OMsg: Msg

The type of messages this actor produces as output.

Required Methods

fn process(&mut self, input: Self::IMsg) -> Vec<Self::OMsg>

Processes an input message and returns a list of output messages.

Arguments

input - The input message to be processed.

Returns

A vector of output messages of type Self::OMsg.