reactor_actor

Trait ActorRecv

pub trait ActorRecv: Send + 'static {
    type IMsg: Msg;

    // Required method
    fn after_recv(
        &mut self,
        worker_id: &str,
        input: &Self::IMsg,
    ) -> impl Future<Output = ChannelAction> + Send;
}

The ActorRecv trait defines what action to take based on incoming message.

It defines a single asynchronous method, ActorRecv::after_recv, that is called after a message is received.

Type Parameters

IMsg: The type of the message this actor receives. It must implement the Msg trait.

Required Associated Types

type IMsg: Msg

The input message type that this actor receives.

Required Methods

fn after_recv( &mut self, worker_id: &str, input: &Self::IMsg, ) -> impl Future<Output = ChannelAction> + Send

Called after the actor receives a message.

This asynchronous method is invoked with:

worker_id: A reference to the address of the sending actor.

input: A reference to the message that was received.

It returns ChannelAction that determines how the actor should proceed.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.